+++ vite.config.js
@@ -1,7 +1,7 @@
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  base: "/front_5th_chapter1-1/",
+  // base: "/front_5th_chapter1-1/",
   test: {
     globals: true,
     environment: "jsdom",
